Refactor: rename to delete new from names of operators in source_lcao
