+++ examples/03_spin_polarized/AFM/reference/running_scf.log.ref
@@ -4811,9 +4811,9 @@ Record_adj           for_2d                       0.00           1   0.00     0.
 Grid_Driver          Find_atom                    0.00         112   0.00     0.01
 LCAO_Hamilt          grid_prepare                 0.00           1   0.00     0.00
 Veff                 initialize_HR                0.00           1   0.00     0.00
-OverlapNew           initialize_SR                0.00           1   0.00     0.00
-EkineticNew          initialize_HR                0.00           1   0.00     0.00
-NonlocalNew          initialize_HR                0.00           1   0.00     0.00
+Overlap              initialize_SR                0.00           1   0.00     0.00
+Ekinetic             initialize_HR                0.00           1   0.00     0.00
+Nonlocal             initialize_HR                0.00           1   0.00     0.00
 Charge               set_rho_core                 0.00           1   0.00     0.00
 Charge               atomic_rho                   0.00           1   0.00     0.01
 PW_Basis_Sup         recip2real                   0.03         219   0.00     0.07
@@ -4838,12 +4838,12 @@ Gint_interface       cal_gint                    27.76          69   0.40    70.
 Gint_interface       cal_gint_vlocal             15.79          46   0.34    40.29
 Gint_Tools           cal_psir_ylm                 2.18        9936   0.00     5.56
 Gint_k               transfer_pvpR                0.75          46   0.02     1.92
-OverlapNew           calculate_SR                 0.09           1   0.09     0.24
-OverlapNew           contributeHk                 0.86        2898   0.00     2.21
-EkineticNew          contributeHR                 0.09          46   0.00     0.24
-EkineticNew          calculate_HR                 0.09           1   0.09     0.24
-NonlocalNew          contributeHR                 0.08          46   0.00     0.20
-NonlocalNew          calculate_HR                 0.07           1   0.07     0.18
+Overlap              calculate_SR                 0.09           1   0.09     0.24
+Overlap              contributeHk                 0.86        2898   0.00     2.21
+Ekinetic             contributeHR                 0.09          46   0.00     0.24
+Ekinetic             calculate_HR                 0.09           1   0.09     0.24
+Nonlocal          contributeHR                 0.08          46   0.00     0.20
+Nonlocal          calculate_HR                 0.07           1   0.07     0.18
 OperatorLCAO         contributeHk                 0.98        2898   0.00     2.49
 HSolverLCAO          hamiltSolvePsiK              6.35        2898   0.00    16.19
 DiagoElpa            elpa_solve                   5.46        2898   0.00    13.93
