+++ examples/03_spin_polarized/ATOM/reference/running_scf.log.ref
@@ -597,9 +597,9 @@ Grid_BigCell         grid_expansion_index         0.01           2   0.01     0.
 Record_adj           for_2d                       0.00           1   0.00     0.00
 Grid_Driver          Find_atom                    0.00           7   0.00     0.00
 LCAO_Hamilt          grid_prepare                 0.00           1   0.00     0.00
-OverlapNew           initialize_SR                0.00           1   0.00     0.00
-EkineticNew          initialize_HR                0.00           1   0.00     0.00
-NonlocalNew          initialize_HR                0.00           1   0.00     0.00
+Overlap              initialize_SR                0.00           1   0.00     0.00
+Ekinetic             initialize_HR                0.00           1   0.00     0.00
+Nonlocal             initialize_HR                0.00           1   0.00     0.00
 Veff                 initialize_HR                0.00           1   0.00     0.00
 LOC                  Alltoall                     0.00           1   0.00     0.00
 Charge               set_rho_core                 0.01           1   0.01     0.03
@@ -621,12 +621,12 @@ H_Ewald_pw           compute_ewald                0.05           1   0.05     0.
 HSolverLCAO          solve                        0.46           7   0.07     2.81
 HamiltLCAO           updateHk                     0.11          14   0.01     0.68
 OperatorLCAO         init                         0.11          42   0.00     0.68
-OverlapNew           calculate_SR                 0.00           1   0.00     0.00
-OverlapNew           contributeHk                 0.00           1   0.00     0.00
-EkineticNew          contributeHR                 0.00          14   0.00     0.00
-EkineticNew          calculate_HR                 0.00           1   0.00     0.00
-NonlocalNew          contributeHR                 0.00          14   0.00     0.00
-NonlocalNew          calculate_HR                 0.00           1   0.00     0.00
+Overlap              calculate_SR                 0.00           1   0.00     0.00
+Overlap              contributeHk                 0.00           1   0.00     0.00
+Ekinetic             contributeHR                 0.00          14   0.00     0.00
+Ekinetic             calculate_HR                 0.00           1   0.00     0.00
+Nonlocal          contributeHR                 0.00          14   0.00     0.00
+Nonlocal          calculate_HR                 0.00           1   0.00     0.00
 Veff                 contributeHR                 0.06           7   0.01     0.39
 Gint_interface       cal_gint                     0.27          23   0.01     1.62
 Gint_interface       cal_gint_vlocal              0.11          14   0.01     0.66
