+++ examples/03_spin_polarized/FM/reference/scf.output.ref
@@ -92,9 +92,9 @@ Record_adj           for_2d                       0.00           1   0.00     0.
 Grid_Driver          Find_atom                    0.00          76   0.00     0.01
 LCAO_Hamilt          grid_prepare                 0.00           1   0.00     0.00
 Veff                 initialize_HR                0.00           1   0.00     0.00
-OverlapNew           initialize_SR                0.00           1   0.00     0.00
-EkineticNew          initialize_HR                0.00           1   0.00     0.00
-NonlocalNew          initialize_HR                0.00           1   0.00     0.01
+Overlap              initialize_SR                0.00           1   0.00     0.00
+Ekinetic             initialize_HR                0.00           1   0.00     0.00
+Nonlocal             initialize_HR                0.00           1   0.00     0.01
 Charge               set_rho_core                 0.00           1   0.00     0.00
 Charge               atomic_rho                   0.00           1   0.00     0.01
 PW_Basis_Sup         recip2real                   0.02         138   0.00     0.07
@@ -119,12 +119,12 @@ Gint_interface       cal_gint                    16.57          42   0.39    68.
 Gint_interface       cal_gint_vlocal              9.47          28   0.34    39.40
 Gint_Tools           cal_psir_ylm                 1.26        6048   0.00     5.23
 Gint_k               transfer_pvpR                0.56          28   0.02     2.32
-OverlapNew           calculate_SR                 0.09           1   0.09     0.39
-OverlapNew           contributeHk                 0.52        1764   0.00     2.18
-EkineticNew          contributeHR                 0.09          28   0.00     0.40
-EkineticNew          calculate_HR                 0.09           1   0.09     0.39
-NonlocalNew          contributeHR                 0.08          28   0.00     0.31
-NonlocalNew          calculate_HR                 0.07           1   0.07     0.29
+Overlap              calculate_SR                 0.09           1   0.09     0.39
+Overlap              contributeHk                 0.52        1764   0.00     2.18
+Ekinetic             contributeHR                 0.09          28   0.00     0.40
+Ekinetic             calculate_HR                 0.09           1   0.09     0.39
+Nonlocal          contributeHR                 0.08          28   0.00     0.31
+Nonlocal          calculate_HR                 0.07           1   0.07     0.29
 OperatorLCAO         contributeHk                 0.59        1764   0.00     2.46
 HSolverLCAO          hamiltSolvePsiK              3.90        1764   0.00    16.25
 DiagoElpa            elpa_solve                   3.33        1764   0.00    13.86
