+++ examples/04_noncollinear/BCC_Fe_NC_ground_state/reference/scf.output.ref
@@ -102,9 +102,9 @@ Record_adj           for_2d                       0.00           1   0.00     0.
 Grid_Driver          Find_atom                    0.00          68   0.00     0.00
 LCAO_Hamilt          grid_prepare                 0.00           1   0.00     0.00
 Veff                 initialize_HR                0.00           1   0.00     0.00
-OverlapNew           initialize_SR                0.00           1   0.00     0.00
-EkineticNew          initialize_HR                0.00           1   0.00     0.00
-NonlocalNew          initialize_HR                0.00           1   0.00     0.00
+Overlap              initialize_SR                0.00           1   0.00     0.00
+Ekinetic             initialize_HR                0.00           1   0.00     0.00
+Nonlocal             initialize_HR                0.00           1   0.00     0.00
 Charge               set_rho_core                 0.01           1   0.01     0.00
 PW_Basis             recip2real                   0.04         306   0.00     0.03
 PW_Basis             gathers_scatterp             0.01         306   0.00     0.01
@@ -129,12 +129,12 @@ Gint_interface       cal_gint                   103.55         120   0.86    90.
 Gint_interface       cal_gint_vlocal             58.66          96   0.61    51.50
 Gint_Tools           cal_psir_ylm                 5.41       17280   0.00     4.75
 Gint_k               transfer_pvpR                4.83          24   0.20     4.24
-OverlapNew           calculate_SR                 0.13           1   0.13     0.11
-OverlapNew           contributeHk                 0.12         192   0.00     0.11
-EkineticNew          contributeHR                 0.13          24   0.01     0.11
-EkineticNew          calculate_HR                 0.13           1   0.13     0.11
-NonlocalNew          contributeHR                 0.79          24   0.03     0.70
-NonlocalNew          calculate_HR                 0.77           1   0.77     0.68
+Overlap              calculate_SR                 0.13           1   0.13     0.11
+Overlap              contributeHk                 0.12         192   0.00     0.11
+Ekinetic             contributeHR                 0.13          24   0.01     0.11
+Ekinetic             calculate_HR                 0.13           1   0.13     0.11
+Nonlocal          contributeHR                 0.79          24   0.03     0.70
+Nonlocal          calculate_HR                 0.77           1   0.77     0.68
 OperatorLCAO         contributeHk                 0.10         192   0.00     0.09
 HSolverLCAO          hamiltSolvePsiK              1.08         192   0.01     0.95
 DiagoElpa            elpa_solve                   0.90         192   0.00     0.79
