+++ source/Makefile.Objects
@@ -339,9 +339,9 @@ OBJS_HAMILT_OF=kedf_tf.o\
 
 OBJS_HAMILT_LCAO=hamilt_lcao.o\
     operator_lcao.o\
-    ekinetic_new.o\
-    nonlocal_new.o\
-    overlap_new.o\
+    ekinetic.o\
+    nonlocal.o\
+    overlap.o\
     td_ekinetic_lcao.o\
     td_nonlocal_lcao.o\
     td_pot_hybrid.o\
