+++ source/source_lcao/CMakeLists.txt
@@ -14,9 +14,9 @@ if(ENABLE_LCAO)
         module_operator_lcao/op_dftu_lcao.cpp
         module_operator_lcao/deepks_lcao.cpp
         module_operator_lcao/op_exx_lcao.cpp
-        module_operator_lcao/overlap_new.cpp
-        module_operator_lcao/ekinetic_new.cpp
-        module_operator_lcao/nonlocal_new.cpp
+        module_operator_lcao/overlap.cpp
+        module_operator_lcao/ekinetic.cpp
+        module_operator_lcao/nonlocal.cpp
         module_operator_lcao/td_ekinetic_lcao.cpp
         module_operator_lcao/td_nonlocal_lcao.cpp
         module_operator_lcao/td_pot_hybrid.cpp
