+++ source/source_lcao/module_operator_lcao/CMakeLists.txt
@@ -6,9 +6,9 @@ add_library(
     meta_lcao.cpp
     veff_lcao.cpp
     deepks_lcao.cpp
-    overlap_new.cpp
-    ekinetic_new.cpp
-    nonlocal_new.cpp
+    overlap.cpp
+    ekinetic.cpp
+    nonlocal.cpp
     td_ekinetic_lcao.cpp
     td_nonlocal_lcao.cpp
     td_pot_hybrid.cpp
