+++ source/source_lcao/module_operator_lcao/ekinetic.h
@@ -1,76 +1,89 @@
-#ifndef NONLOCALNEW_H
-#define NONLOCALNEW_H
+#ifndef EKINETIC_H
+#define EKINETIC_H
 #include "source_basis/module_ao/parallel_orbitals.h"
 #include "source_basis/module_nao/two_center_integrator.h"
 #include "source_cell/module_neighbor/sltk_grid_driver.h"
 #include "source_cell/unitcell.h"
 #include "source_lcao/module_operator_lcao/operator_lcao.h"
 #include "source_lcao/module_hcontainer/hcontainer.h"
-
-#include <unordered_map>
 #include <vector>
 
 namespace hamilt
 {
 
-#ifndef __NONLOCALNEWTEMPLATE
-#define __NONLOCALNEWTEMPLATE
+#ifndef __EKINETICTEMPLATE
+#define __EKINETICTEMPLATE
 
-/// The NonlocalNew class template inherits from class T
-/// it is used to calculate the non-local pseudopotential of wavefunction basis
+/// The EKinetic class template inherits from class T
+/// it is used to calculate the electronic kinetic
 /// Template parameters:
 /// - T: base class, it would be OperatorLCAO<TK> or OperatorPW<TK>
 /// - TR: data type of real space Hamiltonian, it would be double or std::complex<double>
 template <class T>
-class NonlocalNew : public T
+class EKinetic : public T
 {
 };
 
 #endif
 
-/// NonlocalNew class template specialization for OperatorLCAO<TK> base class
-/// It is used to calculate the non-local pseudopotential matrix in real space and fold it to k-space
-/// HR = <psi_{mu, 0}|beta_p1>D_{p1, p2}<beta_p2|psi_{nu, R}>
-/// HK = <psi_{mu, k}|beta_p1>D_{p1, p2}<beta_p2|psi_{nu, k}> = \sum_{R} e^{ikR} HR
+/// EKinetic class template specialization for OperatorLCAO<TK> base class
+/// It is used to calculate the electronic kinetic matrix in real space and fold it to k-space
+/// HR = <psi_{mu, 0}|-\Nabla^2|psi_{nu, R}>
+/// HK = <psi_{mu, k}|-\Nabla^2|psi_{nu, k}> = \sum_{R} e^{ikR} HR
 /// Template parameters:
 /// - TK: data type of k-space Hamiltonian
 /// - TR: data type of real space Hamiltonian
 template <typename TK, typename TR>
-class NonlocalNew<OperatorLCAO<TK, TR>> : public OperatorLCAO<TK, TR>
+class EKinetic<OperatorLCAO<TK, TR>> : public OperatorLCAO<TK, TR>
 {
   public:
-    NonlocalNew<OperatorLCAO<TK, TR>>(HS_Matrix_K<TK>* hsk_in,
+    /**
+     * @brief Construct a new EKinetic object
+     */
+    EKinetic<OperatorLCAO<TK, TR>>(HS_Matrix_K<TK>* hsk_in,
                                       const std::vector<ModuleBase::Vector3<double>>& kvec_d_in,
-                                      hamilt::HContainer<TR>* hR_in,
+                                      HContainer<TR>* hR_in,
                                       const UnitCell* ucell_in,
                                       const std::vector<double>& orb_cutoff,
                                       const Grid_Driver* GridD_in,
                                       const TwoCenterIntegrator* intor);
-    ~NonlocalNew<OperatorLCAO<TK, TR>>();
+
+    /**
+     * @brief Destroy the EKinetic object
+     */
+    ~EKinetic<OperatorLCAO<TK, TR>>();
 
     /**
      * @brief contributeHR() is used to calculate the HR matrix
-     * <phi_{\mu, 0}|beta_p1>D_{p1, p2}<beta_p2|phi_{\nu, R}>
+     * <phi_{\mu, 0}|-\Nabla^2|phi_{\nu, R}>
      */
     virtual void contributeHR() override;
 
+    virtual void set_HR_fixed(void*) override;
+
+    /**
+     * @brief calculate force and stress for kinetic operator
+     * @param cal_force whether to calculate force
+     * @param cal_stress whether to calculate stress
+     * @param dmR density matrix in real space
+     * @param force output force matrix (nat x 3)
+     * @param stress output stress matrix (3 x 3)
+     */
     void cal_force_stress(const bool cal_force,
                           const bool cal_stress,
-                          const HContainer<TR>* dmR,
+                          const HContainer<double>* dmR,
                           ModuleBase::matrix& force,
                           ModuleBase::matrix& stress);
 
-    virtual void set_HR_fixed(void*) override;
-
   private:
     const UnitCell* ucell = nullptr;
-
     std::vector<double> orb_cutoff_;
 
     hamilt::HContainer<TR>* HR_fixed = nullptr;
 
-    // the following variable is introduced temporarily during LCAO refactoring
     const TwoCenterIntegrator* intor_ = nullptr;
+
+    const Grid_Driver* gridD = nullptr;
 
     bool allocated = false;
 
@@ -78,15 +91,14 @@ class NonlocalNew<OperatorLCAO<TK, TR>> : public OperatorLCAO<TK, TR>
 
     /**
      * @brief initialize HR, search the nearest neighbor atoms
-     * HContainer is used to store the non-local pseudopotential matrix with specific <I,J,R> atom-pairs
+     * HContainer is used to store the electronic kinetic matrix with specific <I,J,R> atom-pairs
      * the size of HR will be fixed after initialization
      */
     void initialize_HR(const Grid_Driver* GridD_in);
 
     /**
-     * @brief calculate the non-local pseudopotential matrix with specific <I,J,R> atom-pairs
-     * nearest neighbor atoms don't need to be calculated again
-     * loop the atom-pairs in HR and calculate the non-local pseudopotential matrix
+     * @brief calculate the electronic kinetic matrix with specific <I,J,R> atom-pairs
+     * use the adjs_all to calculate the HR matrix
      */
     void calculate_HR();
 
@@ -95,32 +107,27 @@ class NonlocalNew<OperatorLCAO<TK, TR>> : public OperatorLCAO<TK, TR>
      */
     void cal_HR_IJR(const int& iat1,
                     const int& iat2,
-                    const int& T0,
                     const Parallel_Orbitals* paraV,
-                    const std::unordered_map<int, std::vector<double>>& nlm1_all,
-                    const std::unordered_map<int, std::vector<double>>& nlm2_all,
+                    const ModuleBase::Vector3<double>& dtau,
                     TR* data_pointer);
 
-    const Grid_Driver* gridD = nullptr;
-
     /**
-     * @brief calculate the atomic Force of <I,J,R> atom pair
+     * @brief calculate force contribution for atom pair <I,J,R>
      */
     void cal_force_IJR(const int& iat1,
                        const int& iat2,
-                       const int& T0,
                        const Parallel_Orbitals* paraV,
                        const std::unordered_map<int, std::vector<double>>& nlm1_all,
                        const std::unordered_map<int, std::vector<double>>& nlm2_all,
                        const hamilt::BaseMatrix<TR>* dmR_pointer,
                        double* force1,
                        double* force2);
+
     /**
-     * @brief calculate the Stress of <I,J,R> atom pair
+     * @brief calculate stress contribution for atom pair <I,J,R>
      */
     void cal_stress_IJR(const int& iat1,
                         const int& iat2,
-                        const int& T0,
                         const Parallel_Orbitals* paraV,
                         const std::unordered_map<int, std::vector<double>>& nlm1_all,
                         const std::unordered_map<int, std::vector<double>>& nlm2_all,
@@ -129,6 +136,7 @@ class NonlocalNew<OperatorLCAO<TK, TR>> : public OperatorLCAO<TK, TR>
                         const ModuleBase::Vector3<double>& dis2,
                         double* stress);
 
+    /// @brief exact the nearest neighbor atoms from all adjacent atoms
     std::vector<AdjacentAtomInfo> adjs_all;
 };
