+++ source/source_lcao/module_operator_lcao/test/CMakeLists.txt
@@ -4,7 +4,7 @@ remove_definitions(-DUSE_NEW_TWO_CENTER)
 AddTest(
   TARGET operator_overlap_test
   LIBS parameter ${math_libs} psi base device container
-  SOURCES test_overlapnew.cpp ../overlap_new.cpp ../operator_force_stress_utils.cpp ../../module_hcontainer/func_folding.cpp
+  SOURCES test_overlap.cpp ../overlap.cpp ../operator_force_stress_utils.cpp ../../module_hcontainer/func_folding.cpp
   ../../module_hcontainer/base_matrix.cpp ../../module_hcontainer/hcontainer.cpp ../../module_hcontainer/atom_pair.cpp
   ../../../source_basis/module_ao/parallel_orbitals.cpp
   ../../../source_basis/module_ao/ORB_atomic_lm.cpp
@@ -15,7 +15,7 @@ AddTest(
 AddTest(
   TARGET operator_overlap_serial_test
   LIBS parameter ${math_libs} psi base device container
-  SOURCES test_overlapnew_serial.cpp ../overlap_new.cpp ../operator_force_stress_utils.cpp ../../module_hcontainer/func_folding.cpp
+  SOURCES test_overlap_serial.cpp ../overlap.cpp ../operator_force_stress_utils.cpp ../../module_hcontainer/func_folding.cpp
   ../../module_hcontainer/base_matrix.cpp ../../module_hcontainer/hcontainer.cpp ../../module_hcontainer/atom_pair.cpp
   ../../../source_basis/module_ao/parallel_orbitals.cpp
   ../../../source_basis/module_ao/ORB_atomic_lm.cpp
@@ -26,7 +26,7 @@ AddTest(
 AddTest(
   TARGET operator_overlap_cd_test
   LIBS parameter ${math_libs} psi base device container
-  SOURCES test_overlapnew_cd.cpp ../overlap_new.cpp ../operator_force_stress_utils.cpp ../../module_hcontainer/func_folding.cpp
+  SOURCES test_overlap_cd.cpp ../overlap.cpp ../operator_force_stress_utils.cpp ../../module_hcontainer/func_folding.cpp
   ../../module_hcontainer/base_matrix.cpp ../../module_hcontainer/hcontainer.cpp ../../module_hcontainer/atom_pair.cpp
   ../../../source_basis/module_ao/parallel_orbitals.cpp
   ../../../source_basis/module_ao/ORB_atomic_lm.cpp
@@ -37,7 +37,7 @@ AddTest(
 AddTest(
   TARGET operator_ekinetic_test
   LIBS parameter ${math_libs} psi base device container
-  SOURCES test_ekineticnew.cpp ../ekinetic_new.cpp ../operator_force_stress_utils.cpp ../../module_hcontainer/func_folding.cpp
+  SOURCES test_ekinetic.cpp ../ekinetic.cpp ../operator_force_stress_utils.cpp ../../module_hcontainer/func_folding.cpp
   ../../module_hcontainer/base_matrix.cpp ../../module_hcontainer/hcontainer.cpp ../../module_hcontainer/atom_pair.cpp
   ../../../source_basis/module_ao/parallel_orbitals.cpp
   ../../../source_basis/module_ao/ORB_atomic_lm.cpp
@@ -47,7 +47,7 @@ AddTest(
 AddTest(
   TARGET operator_ekinetic_serial_test
   LIBS parameter ${math_libs} psi base device container
-  SOURCES test_ekineticnew_serial.cpp ../ekinetic_new.cpp ../operator_force_stress_utils.cpp ../../module_hcontainer/func_folding.cpp
+  SOURCES test_ekinetic_serial.cpp ../ekinetic.cpp ../operator_force_stress_utils.cpp ../../module_hcontainer/func_folding.cpp
   ../../module_hcontainer/base_matrix.cpp ../../module_hcontainer/hcontainer.cpp ../../module_hcontainer/atom_pair.cpp
   ../../../source_basis/module_ao/parallel_orbitals.cpp
   ../../../source_basis/module_ao/ORB_atomic_lm.cpp
@@ -57,9 +57,9 @@ AddTest(
 AddTest(
   TARGET operator_nonlocal_test
   LIBS parameter ${math_libs} psi base device container
-  SOURCES test_nonlocalnew.cpp ../nonlocal_new.cpp ../../module_hcontainer/func_folding.cpp 
-  ../../module_hcontainer/base_matrix.cpp ../../module_hcontainer/hcontainer.cpp ../../module_hcontainer/atom_pair.cpp  
-  ../../../source_basis/module_ao/parallel_orbitals.cpp 
+  SOURCES test_nonlocal.cpp ../nonlocal.cpp ../operator_force_stress_utils.cpp ../../module_hcontainer/func_folding.cpp
+  ../../module_hcontainer/base_matrix.cpp ../../module_hcontainer/hcontainer.cpp ../../module_hcontainer/atom_pair.cpp
+  ../../../source_basis/module_ao/parallel_orbitals.cpp
   ../../../source_basis/module_ao/ORB_atomic_lm.cpp
   tmp_mocks.cpp ../../../source_hamilt/operator.cpp
 )
@@ -67,7 +67,7 @@ AddTest(
 AddTest(
   TARGET operator_T_NL_cd_test
   LIBS parameter ${math_libs} psi base device container
-  SOURCES test_T_NL_cd.cpp ../nonlocal_new.cpp ../ekinetic_new.cpp ../operator_force_stress_utils.cpp ../../module_hcontainer/func_folding.cpp
+  SOURCES test_T_NL_cd.cpp ../nonlocal.cpp ../ekinetic.cpp ../operator_force_stress_utils.cpp ../../module_hcontainer/func_folding.cpp
   ../../module_hcontainer/base_matrix.cpp ../../module_hcontainer/hcontainer.cpp ../../module_hcontainer/atom_pair.cpp
   ../../../source_basis/module_ao/parallel_orbitals.cpp
   ../../../source_basis/module_ao/ORB_atomic_lm.cpp
