+++ source/source_lcao/module_operator_lcao/test/test_ekinetic.cpp
@@ -1,24 +1,24 @@
-#include "../overlap_new.h"
+#include "../ekinetic.h"
 
 #include "gtest/gtest.h"
 
 //---------------------------------------
-// Unit test of OverlapNew class
-// OverlapNew is a derivative class of Operator, it is used to calculate the overlap matrix
-// It use HContainer to store the real space SR matrix
-// In this test, we test the correctness and time consuming of 3 functions in OverlapNew class
-// - initialize_SR() called in constructor
+// Unit test of EKinetic class
+// EKinetic is a derivative class of Operator, it is used to calculate the kinetic matrix
+// It use HContainer to store the real space HR matrix
+// In this test, we test the correctness and time consuming of 3 functions in EKinetic class
+// - initialize_HR() called in constructor
 // - contributeHR()
 // - contributeHk()
-// - SR(double) and SK(complex<double>) are tested in constructHRd2cd
-// - SR(double) and SK(double) are tested in constructHRd2d
+// - HR(double) and SK(complex<double>) are tested in constructHRd2cd
+// - HR(double) and SK(double) are tested in constructHRd2d
 //---------------------------------------
 
 // test_size is the number of atoms in the unitcell
 // modify test_size to test different size of unitcell
 int test_size = 10;
 int test_nw = 10;
-class OverlapNewTest : public ::testing::Test
+class EKineticTest : public ::testing::Test
 {
   protected:
     void SetUp() override
@@ -58,12 +58,12 @@ class OverlapNewTest : public ::testing::Test
         ucell.set_iat2iwt(1);
         init_parav();
         // set up a HContainer with ucell
-        SR = new hamilt::HContainer<double>(paraV);
+        HR = new hamilt::HContainer<double>(paraV);
     }
 
     void TearDown() override
     {
-        delete SR;
+        delete HR;
         delete paraV;
         delete[] ucell.atoms;
     }
@@ -86,7 +86,7 @@ class OverlapNewTest : public ::testing::Test
 #endif
 
     UnitCell ucell;
-    hamilt::HContainer<double>* SR;
+    hamilt::HContainer<double>* HR;
     Parallel_Orbitals* paraV;
     TwoCenterIntegrator intor_;
 
@@ -94,20 +94,20 @@ class OverlapNewTest : public ::testing::Test
     int my_rank = 0;
 };
 
-// using TEST_F to test OverlapNew
-TEST_F(OverlapNewTest, constructHRd2d)
+// using TEST_F to test EKinetic
+TEST_F(EKineticTest, constructHRd2d)
 {
     std::vector<ModuleBase::Vector3<double>> kvec_d_in(1, ModuleBase::Vector3<double>(0.0, 0.0, 0.0));
-    hamilt::HS_Matrix_K<double> hsk(paraV);
-    hsk.set_zero_sk();
+    hamilt::HS_Matrix_K<double> hsk(paraV, true);
+    hsk.set_zero_hk();
     Grid_Driver gd(0, 0);
-    hamilt::OverlapNew<hamilt::OperatorLCAO<double, double>>
-        op(&hsk, kvec_d_in, nullptr, SR, &ucell, {1.0}, &gd, &intor_);
+    hamilt::EKinetic<hamilt::OperatorLCAO<double, double>>
+        op(&hsk, kvec_d_in, HR, &ucell, {1.0}, &gd, &intor_);
     op.contributeHR();
-    // check the value of SR
-    for (int iap = 0; iap < SR->size_atom_pairs(); ++iap)
+    // check the value of HR
+    for (int iap = 0; iap < HR->size_atom_pairs(); ++iap)
     {
-        hamilt::AtomPair<double>& tmp = SR->get_atom_pair(iap);
+        hamilt::AtomPair<double>& tmp = HR->get_atom_pair(iap);
         int iat1 = tmp.get_atom_i();
         int iat2 = tmp.get_atom_j();
         auto indexes1 = paraV->get_indexes_row(iat1);
@@ -118,30 +118,46 @@ TEST_F(OverlapNewTest, constructHRd2d)
             EXPECT_EQ(tmp.get_pointer(0)[i], 1.0);
         }
     }
-    // calculate SK
+    // calculate HK
     op.contributeHk(0);
-    // check the value of SK
-    double* sk = hsk.get_sk();
-    for (int i = 0; i < hsk.get_size(); ++i)
+    // check the value of HK
+    double* hk = hsk.get_hk();
+    for (int i = 0; i < paraV->get_row_size() * paraV->get_col_size(); ++i)
     {
-        EXPECT_EQ(sk[i], 1.0);
+        EXPECT_EQ(hk[i], 1.0);
+    }
+    // calculate HR again
+    op.contributeHR();
+    // check the value of HR
+    for (int iap = 0; iap < HR->size_atom_pairs(); ++iap)
+    {
+        hamilt::AtomPair<double>& tmp = HR->get_atom_pair(iap);
+        int iat1 = tmp.get_atom_i();
+        int iat2 = tmp.get_atom_j();
+        auto indexes1 = paraV->get_indexes_row(iat1);
+        auto indexes2 = paraV->get_indexes_col(iat2);
+        int nwt = indexes1.size() * indexes2.size();
+        for (int i = 0; i < nwt; ++i)
+        {
+            EXPECT_EQ(tmp.get_pointer(0)[i], 2.0);
+        }
     }
 }
 
-TEST_F(OverlapNewTest, constructHRd2cd)
+TEST_F(EKineticTest, constructHRd2cd)
 {
     std::vector<ModuleBase::Vector3<double>> kvec_d_in(2, ModuleBase::Vector3<double>(0.0, 0.0, 0.0));
     kvec_d_in[1] = ModuleBase::Vector3<double>(0.1, 0.2, 0.3);
-    hamilt::HS_Matrix_K<std::complex<double>> hsk(paraV);
-    hsk.set_zero_sk();
+    hamilt::HS_Matrix_K<std::complex<double>> hsk(paraV, true);
+    hsk.set_zero_hk();
     Grid_Driver gd(0, 0);
-    hamilt::OverlapNew<hamilt::OperatorLCAO<std::complex<double>, double>>
-        op(&hsk, kvec_d_in, nullptr, SR, &ucell, {1.0}, &gd, &intor_);
+    hamilt::EKinetic<hamilt::OperatorLCAO<std::complex<double>, double>>
+        op(&hsk, kvec_d_in, HR, &ucell, {1.0}, &gd, &intor_);
     op.contributeHR();
-    // check the value of SR
-    for (int iap = 0; iap < SR->size_atom_pairs(); ++iap)
+    // check the value of HR
+    for (int iap = 0; iap < HR->size_atom_pairs(); ++iap)
     {
-        hamilt::AtomPair<double>& tmp = SR->get_atom_pair(iap);
+        hamilt::AtomPair<double>& tmp = HR->get_atom_pair(iap);
         int iat1 = tmp.get_atom_i();
         int iat2 = tmp.get_atom_j();
         auto indexes1 = paraV->get_indexes_row(iat1);
@@ -152,23 +168,23 @@ TEST_F(OverlapNewTest, constructHRd2cd)
             EXPECT_EQ(tmp.get_pointer(0)[i], 1.0);
         }
     }
-    // calculate SK for gamma point
+    // calculate HK for gamma point
     op.contributeHk(0);
-    // check the value of SK of gamma point
-    auto* sk = hsk.get_sk();
+    auto* hk = hsk.get_hk();
+    // check the value of HK of gamma point
     for (int i = 0; i < paraV->get_row_size() * paraV->get_col_size(); ++i)
     {
-        EXPECT_EQ(sk[i].real(), 1.0);
-        EXPECT_EQ(sk[i].imag(), 0.0);
+        EXPECT_EQ(hk[i].real(), 1.0);
+        EXPECT_EQ(hk[i].imag(), 0.0);
     }
-    // calculate SK for k point
-    hsk.set_zero_sk();
+    // calculate HK for k point
+    hsk.set_zero_hk();
     op.contributeHk(1);
-    // check the value of SK
+    // check the value of HK
     for (int i = 0; i < paraV->get_row_size() * paraV->get_col_size(); ++i)
     {
-        EXPECT_NEAR(sk[i].real(), -0.80901699437494723, 1e-10);
-        EXPECT_NEAR(sk[i].imag(), -0.58778525229247336, 1e-10);
+        EXPECT_NEAR(hk[i].real(), -1.6180339887498945 / 2, 1e-10);
+        EXPECT_NEAR(hk[i].imag(), -1.1755705045849467 / 2, 1e-10);
     }
 }
